Change the definition of $qty_base
$qty_base is the baseline value of $qty before a spate. It was previously defined as:

$qty_base = $last_qty_vals[0];

which is the value of $qty at point $ndata-($n_recent_to_consider-1). This value is in fact the $qty value at the point where the raise is first observed (see in_spate condition). It was therefore changed to

$qty_base = $dq_last_vals[0];

which is the $qty value at point $ndata-$n_recent_to_consider - i.e. just before the raise was first observed.

TODO: use an average value on several points before the spate