Further modif of .dat file + correction of BUG9
The recent additions in the .dat file behaves now as follow:
	o qty_delta: prints difference between qty_base (qty at
		point just before first observed raise) and qty_max
		(qty at peak) at point just after peak was detected
	o qty_delta_dn: prints difference between qty_base (qty at
		point just before first observed raise) and qty_min
		(qty at through) at point just after through was
		detected
	o qty_max: prints qty value at peak just after peak was
		detected
	o qty_min: prints qty value at through just after through
		was detected
	o qty_base: prints qty_base (value at point just before
		first raise is observed) at point where a new
		spate is detected
	o spate_start: prints the index of the point just before
		first raise was observed at point where a new
		spate is detected
Two additional variables are now printed:
	o is_peak: prints the index of the point where a peak was
		detected at following point
	o is_through: prints the index of the point where a through
		was detected at following point
Ideally, one spate should be defined by only one peak and one
through. Thought was given on how to implement this in the code. In
hydrology, generally, the highest local max is considered as the
spate peak. This is not straightforward to implement in the code,
as the script has no knowledge on what happens next - i.e. if a
greater peak will be observed in the current spate. Furthermore,
the script was not intended to work this way, but rather to detect
spate starts, peaks and throughs relying only on what happened
before. Keeping that in mind, it seems easier to select a unique
peak and through per spate in post-processing, on the basis of the
info currently given in the .dat file.

BUG9: identified by JF 160418 - the detection of a through was
awkwardly defined due to its inheritance from the peak detection
(one sign error and an incorrect variable name have been
corrected).